USBMonitor: don't query for device's details
The 'newAPI' approach for obtaining device data (for creating a hash map
for all enumerated device) was asking for some detailed information
(such as serial number) which was causing Android exceptions. The reason
is that this query was issued before giving the permissions by a user to
access an ecternal USB device.